refactor(recipes): move recipe key helper functions to utils module
