+++ src/controllers/bookmarks/helper.ts
@@ -1,5 +1,5 @@
 import { BookmarkRecipe } from "../../router/v1/bookmarks/index.js";
-import { generateRecipeKey } from "../../services/recipes/helper.js";
+import { generateRecipeKey } from "../../services/recipes/utils.js";
 
 export const generateRecipeKeysForBookmarks = (bookmarks: BookmarkRecipe[]) =>
   bookmarks.map((bookmark) =>
