+++ src/controllers/recipes/recipe/deleteRecipeController.ts
@@ -1,6 +1,6 @@
 import { NextFunction, Request, Response } from "express";
 
-import { decryptRecipeURLAndGetRecipeId } from "../../../services/recipes/recipe/helper.js";
+import { decryptRecipeURLAndGetRecipeId } from "../../../services/recipes/utils.js";
 import { validateId } from "../../../utils/numbers.js";
 import { ServiceError } from "../../../services/helpers/ServiceError.js";
 import {
