+++ src/services/recipes/recipesCreateService/index.ts
@@ -5,7 +5,7 @@ import mysqlDB from "../../../db/mysql.js";
 import { SerializedUser } from "../../../config/passport.js";
 
 import { INewRecipe } from "../type.js";
-import { generateRecipeKey } from "../helper.js";
+import { generateRecipeKey } from "../utils.js";
 import { CreateRecipeBody } from "../../../controllers/recipes/recipesCreateController.js";
 import {
   insertIngredients,
